Re-factors Material class so that ANY material may become 'discrete'
(refs #6533)
