+++ framework/include/actions/MaterialOutputAction.h
@@ -72,7 +72,7 @@ private:
    * act() method.
    */
   template<typename T>
-  void materialOutputHelper(const std::string & property_name, MooseSharedPointer<MaterialBase> material);
+  void materialOutputHelper(const std::string & property_name, MooseSharedPointer<Material> material);
 
   /**
    * A method for creating an AuxVariable and associated action
@@ -82,7 +82,7 @@ private:
    * @param material A pointer to the Material object containing the property of interest
    */
   MooseSharedPointer<MooseObjectAction> createAction(const std::string & type, const std::string & property_name,
-                                                     const std::string & variable_name, MooseSharedPointer<MaterialBase> material);
+                                                     const std::string & variable_name, MooseSharedPointer<Material> material);
 
   /// Pointer the MaterialData object storing the block restricted materials
   MooseSharedPointer<MaterialData> _block_material_data;
@@ -109,7 +109,7 @@ private:
 
 template<typename T>
 void
-MaterialOutputAction::materialOutputHelper(const std::string & /*property_name*/, MooseSharedPointer<MaterialBase> /*material*/)
+MaterialOutputAction::materialOutputHelper(const std::string & /*property_name*/, MooseSharedPointer<Material> /*material*/)
 {
   mooseError("Unknown type, you must create a specialization of materialOutputHelper");
 }
