+++ framework/include/materials/Material.h
@@ -15,34 +15,384 @@
 #ifndef MATERIAL_H
 #define MATERIAL_H
 
-#include "MaterialBase.h"
+// MOOOSE includes
+#include "MooseObject.h"
+#include "BlockRestrictable.h"
+#include "BoundaryRestrictable.h"
+#include "SetupInterface.h"
+#include "Coupleable.h"
+#include "MooseVariableDependencyInterface.h"
+#include "ScalarCoupleable.h"
+#include "FunctionInterface.h"
+#include "UserObjectInterface.h"
+#include "TransientInterface.h"
+#include "MaterialPropertyInterface.h"
+#include "PostprocessorInterface.h"
+#include "DependencyResolverInterface.h"
+#include "Restartable.h"
+#include "ZeroInterface.h"
+#include "MeshChangedInterface.h"
+#include "OutputInterface.h"
+#include "RandomInterface.h"
+#include "MaterialProperty.h"
 
-// Forward declarations
+// forward declarations
 class Material;
+class MooseMesh;
+class MaterialData;
+class SubProblem;
 
 template<>
 InputParameters validParams<Material>();
 
 /**
- * Holds material properties that are assigned to blocks and computed by MOOSE.
+ * Materials compute MaterialProperties.
  */
-class Material : public MaterialBase
+class Material :
+    public MooseObject,
+    public BlockRestrictable,
+    public BoundaryRestrictable,
+    public SetupInterface,
+    public Coupleable,
+    public MooseVariableDependencyInterface,
+    public ScalarCoupleable,
+    public FunctionInterface,
+    public UserObjectInterface,
+    public TransientInterface,
+    public MaterialPropertyInterface,
+    public PostprocessorInterface,
+    public DependencyResolverInterface,
+    public Restartable,
+    public ZeroInterface,
+    public MeshChangedInterface,
+    public OutputInterface,
+    public RandomInterface
 {
 public:
 
+  /**
+   * Constructor
+   */
   Material(const InputParameters & parameters);
+
+  /**
+   * Initialize stateful properties (if material has some)
+   */
+  virtual void initStatefulProperties(unsigned int n_points);
 
   /**
    * Performs the quadrature point loop, calling computeQpProperties
    */
   virtual void computeProperties();
 
+  /**
+   * Resets the properties at each quadrature point (see resetQpProperties), only called if 'compute = false'.
+   *
+   * This method is called internally by MOOSE, you probably don't want to mess with this.
+   */
+  virtual void resetProperties();
+
+  /**
+   * A method for (re)computing the properties of a Material.
+   *
+   * This is intended to be called from other objects, by first calling
+   * MaterialPropertyInterface::getMaterial and then calling this method on the Material object returned.
+   */
+  virtual void computeProperties(unsigned int qp);
+
+  ///@{
+  /**
+   * Retrieve the property through a given input parameter key with a fallback
+   * to getting it by name
+   */
+  template<typename T>
+  const MaterialProperty<T> & getMaterialProperty(const std::string & name);
+  template<typename T>
+  const MaterialProperty<T> & getMaterialPropertyOld(const std::string & name);
+  template<typename T>
+  const MaterialProperty<T> & getMaterialPropertyOlder(const std::string & name);
+  ///@}
+
+  ///@{
+  /**
+   * Retrieve the property named "name"
+   */
+  template<typename T>
+  const MaterialProperty<T> & getMaterialPropertyByName(const std::string & prop_name);
+  template<typename T>
+  const MaterialProperty<T> & getMaterialPropertyOldByName(const std::string & prop_name);
+  template<typename T>
+  const MaterialProperty<T> & getMaterialPropertyOlderByName(const std::string & prop_name);
+  ///@}
+
+  ///@{
+  /**
+   * Declare the property named "name"
+   */
+  template<typename T>
+  MaterialProperty<T> & declareProperty(const std::string & prop_name);
+  template<typename T>
+  MaterialProperty<T> & declarePropertyOld(const std::string & prop_name);
+  template<typename T>
+  MaterialProperty<T> & declarePropertyOlder(const std::string & prop_name);
+  ///@}
+
+  /**
+   * Return a material property that is initialized to zero by default and does
+   * not need to (but can) be declared by another material.
+   */
+  template<typename T>
+  const MaterialProperty<T> & getZeroMaterialProperty(const std::string & prop_name);
+
+  /**
+   * Return a set of properties accessed with getMaterialProperty
+   * @return A reference to the set of properties with calls to getMaterialProperty
+   */
+  virtual
+  const std::set<std::string> &
+  getRequestedItems() { return _requested_props; }
+
+  /**
+   * Return a set of properties accessed with declareProperty
+   * @return A reference to the set of properties with calls to declareProperty
+   */
+  virtual
+  const std::set<std::string> &
+  getSuppliedItems() { return _supplied_props; }
+
+  void checkStatefulSanity() const;
+
+  /**
+   * Check if a material property is valid for all blocks of this Material
+   *
+   * This method returns true if the supplied property name has been declared
+   * in a Material object on the block ids for this object.
+   *
+   * @param prop_name the name of the property to query
+   * @return true if the property exists for all block ids of the object, otherwise false
+   *
+   * @see BlockRestrictable::hasBlockMaterialPropertyHelper
+   */
+  virtual bool hasBlockMaterialPropertyHelper(const std::string & prop_name);
+
+  /**
+   * Get the list of output objects that this class is restricted
+   * @return A vector of OutputNames
+   */
+  std::set<OutputName> getOutputs();
+
+  /**
+   * Returns true of the MaterialData type is not associated with volume data
+   */
+  bool isBoundaryMaterial() { return _bnd; }
+
 protected:
+
   /**
    * Users must override this method.
    */
   virtual void computeQpProperties();
+
+  /**
+   * Resets the properties prior to calculation of traditional materials (only if 'compute = false').
+   *
+   * This method must be overridden in your class. This is called just prior to the re-calculation of
+   * traditional material properties to ensure that the properties are in a proper state for calculation.
+   */
+  virtual void resetQpProperties();
+
+  /**
+   * Initialize stateful properties at quadrature points.  Note when using this function you only need to address
+   * the "current" material properties not the old ones directly, i.e. if you have a property named "_diffusivity"
+   * and an older property named "_diffusivity_old".  You only need to initialize diffusivity.  MOOSE will use
+   * copy that initial value to the old and older values as necessary.
+   */
+  virtual void initQpStatefulProperties();
+
+  SubProblem & _subproblem;
+
+  FEProblem & _fe_problem;
+  THREAD_ID _tid;
+  Assembly & _assembly;
+
+  bool _bnd;
+  bool _neighbor;
+
+  unsigned int _qp;
+
+  QBase * & _qrule;
+  const MooseArray<Real> & _JxW;
+  const MooseArray<Real> & _coord;
+  const MooseArray<Point> & _q_point;
+  /// normals at quadrature points (valid only in boundary materials)
+  const MooseArray<Point> & _normals;
+
+  const Elem * & _current_elem;
+
+  /// current side of the current element
+  unsigned int & _current_side;
+
+  MooseMesh & _mesh;
+
+  /// Coordinate system
+  const Moose::CoordinateSystemType & _coord_sys;
+
+  /// Set of properties accessed via get method
+  std::set<std::string> _requested_props;
+
+  /// Set of properties declared
+  std::set<std::string> _supplied_props;
+
+  /// If False MOOSE does not compute this property
+  bool _compute;
+
+  enum QP_Data_Type {
+    CURR,
+    PREV
+  };
+
+  enum Prop_State
+  {
+    CURRENT = 0x1,
+    OLD = 0x2,
+    OLDER = 0x4
+  };
+  std::map<std::string, int> _props_to_flags;
+
+
+private:
+  /**
+   * Small helper function to call storeMatPropName
+   */
+  void registerPropName(std::string prop_name, bool is_get, Prop_State state);
+
+  bool _has_stateful_property;
 };
 
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialProperty(const std::string & name)
+{
+  // Check if the supplied parameter is a valid imput parameter key
+  std::string prop_name = deducePropertyName(name);
 
-#endif //MATERIAL_H
+  // Check if it's just a constant.
+  const MaterialProperty<T> * default_property = defaultMaterialProperty<T>(prop_name);
+  if (default_property)
+    return *default_property;
+
+  return getMaterialPropertyByName<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialPropertyOld(const std::string & name)
+{
+  // Check if the supplied parameter is a valid imput parameter key
+  std::string prop_name = deducePropertyName(name);
+
+  // Check if it's just a constant.
+  const MaterialProperty<T> * default_property = defaultMaterialProperty<T>(prop_name);
+  if (default_property)
+    return *default_property;
+
+  return getMaterialPropertyOldByName<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialPropertyOlder(const std::string & name)
+{
+  // Check if the supplied parameter is a valid imput parameter key
+  std::string prop_name = deducePropertyName(name);
+
+  // Check if it's just a constant.
+  const MaterialProperty<T> * default_property = defaultMaterialProperty<T>(prop_name);
+  if (default_property)
+    return *default_property;
+
+  return getMaterialPropertyOlderByName<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialPropertyByName(const std::string & prop_name)
+{
+  // The property may not exist yet, so declare it (declare/getMaterialProperty are referencing the same memory)
+  _requested_props.insert(prop_name);
+  registerPropName(prop_name, true, Material::CURRENT);
+  _fe_problem.markMatPropRequested(prop_name);
+  return _material_data->getProperty<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialPropertyOldByName(const std::string & prop_name)
+{
+  _requested_props.insert(prop_name);
+  registerPropName(prop_name, true, Material::OLD);
+  _fe_problem.markMatPropRequested(prop_name);
+  return _material_data->getPropertyOld<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getMaterialPropertyOlderByName(const std::string & prop_name)
+{
+  _requested_props.insert(prop_name);
+  registerPropName(prop_name, true, Material::OLDER);
+  _fe_problem.markMatPropRequested(prop_name);
+  return _material_data->getPropertyOlder<T>(prop_name);
+}
+
+
+template<typename T>
+MaterialProperty<T> &
+Material::declareProperty(const std::string & prop_name)
+{
+  registerPropName(prop_name, false, Material::CURRENT);
+  return _material_data->declareProperty<T>(prop_name);
+}
+
+template<typename T>
+MaterialProperty<T> &
+Material::declarePropertyOld(const std::string & prop_name)
+{
+  registerPropName(prop_name, false, Material::OLD);
+  return _material_data->declarePropertyOld<T>(prop_name);
+}
+
+template<typename T>
+MaterialProperty<T> &
+Material::declarePropertyOlder(const std::string & prop_name)
+{
+  registerPropName(prop_name, false, Material::OLDER);
+  return _material_data->declarePropertyOlder<T>(prop_name);
+}
+
+template<typename T>
+const MaterialProperty<T> &
+Material::getZeroMaterialProperty(const std::string & prop_name)
+{
+  MaterialProperty<T> & preload_with_zero = _material_data->getProperty<T>(prop_name);
+
+  _requested_props.insert(prop_name);
+  registerPropName(prop_name, true, Material::CURRENT);
+  _fe_problem.markMatPropRequested(prop_name);
+
+  // Register this material on these blocks and boundaries as a zero property with relaxed consistency checking
+  for (std::set<SubdomainID>::const_iterator it = blockIDs().begin(); it != blockIDs().end(); ++it)
+    _fe_problem.storeZeroMatProp(*it, prop_name);
+  for (std::set<BoundaryID>::const_iterator it = boundaryIDs().begin(); it != boundaryIDs().end(); ++it)
+    _fe_problem.storeZeroMatProp(*it, prop_name);
+
+  // set values for all qpoints to zero
+  unsigned int nqp = _mi_feproblem.getMaxQps();
+  preload_with_zero.resize(nqp);
+  for (unsigned int qp = 0; qp < nqp; ++qp)
+    mooseSetToZero<T>(preload_with_zero[qp]);
+
+  return preload_with_zero;
+}
+
+#endif // MATERIAL_H
