+++ framework/include/materials/MaterialData.h
@@ -25,7 +25,6 @@
 #include <vector>
 
 class Material;
-class DiscreteMaterial;
 
 /**
  * Proxy for accessing MaterialPropertyStorage.
@@ -86,8 +85,8 @@ public:
   // Reinit material properties for given element (and possible side)
   void reinit(const std::vector<MooseSharedPointer<Material> > & mats);
 
-  /// Calls the reset method of DiscreteMaterials to ensure that they are in a proper state.
-  void reset(const std::vector<MooseSharedPointer<DiscreteMaterial> > & mats);
+  /// Calls the reset method of Materials to ensure that they are in a proper state.
+  void reset(const std::vector<MooseSharedPointer<Material> > & mats);
 
   // material properties for given element (and possible side)
   void swapBack(const Elem & elem, unsigned int side = 0);
