+++ framework/include/materials/MaterialPropertyStorage.h
@@ -20,7 +20,7 @@
 #include "HashMap.h"
 
 // Forward declarations
-class MaterialBase;
+class Material;
 class MaterialData;
 class QpMap;
 
@@ -108,7 +108,7 @@ public:
    * @param elem Element we are on
    * @param side Side of the element 'elem' (0 for volumetric material properties)
    */
-  void initStatefulProps(MaterialData & material_data, const std::vector<MooseSharedPointer<MaterialBase> > & mats, unsigned int n_qpoints, const Elem & elem, unsigned int side = 0);
+  void initStatefulProps(MaterialData & material_data, const std::vector<MooseSharedPointer<Material> > & mats, unsigned int n_qpoints, const Elem & elem, unsigned int side = 0);
 
   /**
    * Shift the material properties in time.
