+++ modules/solid_mechanics/src/materials/SolidModel.C
@@ -859,7 +859,7 @@ SolidModel::initialSetup()
   {
 
 //    const std::vector<Material*> * mats_p;
-    std::vector<MooseSharedPointer<MaterialBase> > const * mats_p;
+    std::vector<MooseSharedPointer<Material> > const * mats_p;
     std::string suffix;
     if (_bnd)
     {
@@ -870,7 +870,7 @@ SolidModel::initialSetup()
       mats_p = &_fe_problem.getMaterialWarehouse().getActiveBlockObjects(_block_id[i], _tid);
 
 
-    const std::vector<MooseSharedPointer<MaterialBase> > & mats = *mats_p;
+    const std::vector<MooseSharedPointer<Material> > & mats = *mats_p;
 
     for (unsigned int j=0; j < mats.size(); ++j)
     {
