+++ test/src/materials/NewtonMaterial.C
@@ -14,7 +14,7 @@
 
 // MOOSE includes
 #include "NewtonMaterial.h"
-#include "DiscreteMaterial.h"
+#include "Material.h"
 
 
 template<>
@@ -23,10 +23,10 @@ InputParameters validParams<NewtonMaterial>()
   InputParameters params = validParams<Material>();
   params.addRequiredParam<std::string>("f_name", "The name of the property that holds the value of the function for which the root is being computed");
   params.addRequiredParam<std::string>("f_prime_name", "The name of the property that holds the value to of the derivative of the function");
-  params.addRequiredParam<std::string>("p_name", "The name of the independant variable for the function");
+  params.addRequiredParam<std::string>("p_name", "The name of the independent variable for the function");
   params.addParam<Real>("tol", 1e-12, "Newton solution tolerance.");
   params.addParam<unsigned int>("max_iterations", 42, "The maximum number of Newton iterations.");
-  params.addRequiredParam<DiscreteMaterialName>("material", "The material object to recompute.");
+  params.addRequiredParam<MaterialName>("material", "The material object to recompute.");
   return params;
 }
 
@@ -37,7 +37,7 @@ NewtonMaterial::NewtonMaterial(const InputParameters & parameters) :
     _f_prime(getMaterialProperty<Real>(getParam<std::string>("f_prime_name"))),
     _p(declareProperty<Real>(getParam<std::string>("p_name"))),
     _max_iterations(getParam<unsigned int>("max_iterations")),
-    _discrete(getDiscreteMaterial("material"))
+    _discrete(getMaterial("material"))
 {
 }
